Moved user role to make new FE function work.
