+++ webapp/app/common/user-role.ts
@@ -1,0 +1,3 @@
+import { z } from "zod";
+
+export const UserRole = z.enum(["guardian", "securityCouncil", "zkFoundation", "visitor"]);
